chore(pubspec): update blockchain_utils dependency to use Git source and add publish_to field
This commit modifies the pubspec.yaml file by commenting out the previous blockchain_utils dependency version and replacing it with a Git source reference. Additionally, a publish_to field is added, indicating that the package is not intended for publishing. These changes streamline dependency management and clarify the package's publishing status.